+++ lib/pages/admin_exercise_add_screen.dart
@@ -76,29 +76,25 @@ class _AdminExerciseAddScreenState extends State<AdminExerciseAddScreen> {
                 'Save Exercise',
                 onPressed: () async {
                   if (!nameBloc.isValid()) {
-                    return Toast.info(
+                    return ToastController().show(
                       nameBloc.errorText!,
-                      context: context,
                     );
                   }
                   if (!descriptionBloc.isValid()) {
-                    return Toast.info(
+                    return ToastController().show(
                       descriptionBloc.errorText!,
-                      context: context,
                     );
                   }
 
                   if (musclesController.state.isEmpty) {
-                    return Toast.info(
+                    return ToastController().show(
                       'Please select at least one muscle',
-                      context: context,
                     );
                   }
 
                   if (imageFile == null) {
-                    return Toast.info(
+                    return ToastController().show(
                       'Please select an image',
-                      context: context,
                     );
                   }
                   // generate id, for storage and firestore
@@ -128,9 +124,8 @@ class _AdminExerciseAddScreenState extends State<AdminExerciseAddScreen> {
                     );
                   }).catchError(
                     (e) {
-                      Toast.info(
+                      ToastController().show(
                         'Error ${widget.exercise == null ? 'adding' : 'updating'} exercise: $e',
-                        context: context,
                       );
                     },
                   );