+++ lib/pages/user_exercise_delete_popup.dart
@@ -38,7 +38,7 @@ class UserExerciseDeletePopup extends StatelessWidget {
                 );
                 await UserRepository.deleteUsersExercise(widget.exercise!);
               } catch (e) {
-                Toast.info(e, context: context);
+                ToastController().show(e);
                 Navigation.pop();
                 return;
               }