+++ lib/pages/user_workout_add_screen.dart
@@ -207,15 +207,13 @@ class _UserWorkoutAddScreenState extends State<UserWorkoutAddScreen> {
           widget.workout != null ? 'Save Workout' : 'Add Workout',
           onPressed: () async {
             if (!nameBloc.isValid()) {
-              return Toast.info(
+              return ToastController().show(
                 nameBloc.errorText,
-                context: context,
               );
             }
             if (!descriptionBloc.isValid()) {
-              return Toast.info(
+              return ToastController().show(
                 descriptionBloc.errorText ?? 'Invalid description',
-                context: context,
               );
             }
             try {
@@ -240,7 +238,7 @@ class _UserWorkoutAddScreenState extends State<UserWorkoutAddScreen> {
               }
               Navigation.flush(widget: const UserHomeScreen(initialIndex: 1));
             } catch (e) {
-              Toast.info(e, context: context);
+              ToastController().show(e);
             }
           },
         ),